requirements.txt
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
blis==0.4.1
Bottleneck==1.3.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
cymem==2.0.3
dataclasses==0.6
decorator==4.4.1
defusedxml==0.6.0
dnspython==1.16.0
entrypoints==0.3
eventlet==0.25.1
fastai==1.0.60
fastprogress==0.2.2
glob2==0.7
greenlet==0.4.15
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
monotonic==1.5
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.6
notebook==6.0.3
numexpr==2.7.1
numpy==1.18.1
nvidia-ml-py3==7.352.0
opencv-python==4.2.0.32
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
soupsieve==1.9.5
spacy==2.2.3
srsly==1.0.1
terminado==0.8.3
testpath==0.4.4
thinc==7.3.1
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==2.2.0
